pyproject.toml
[project]
name = "tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.121.3",
    "jinja2>=3.1.6",
    "markdown>=3.10",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.44",
    "taskipy>=1.14.1",
    "uvicorn>=0.38.0",
]

[tool.taskipy.tasks]
migrate = "alembic upgrade head"
dev = "uv run uvicorn app.main:app --host 127.0.0.1 --port 8000 --reload"
prod_migrate = "uv run --env-file prod.env alembic upgrade head"
prod_start = "uv run --env-file prod.env uvicorn app.main:app --host 0.0.0.0 --port 4000 --workers 1"

# Main Production Command
prod = "task prod_migrate && task prod_start"